Chore#66 category 패키지 루트로 이동
* fix: WebMvcConfig 중복 제거

* chore: category 패키지 루트로 이동

* test: DisplayName 잘못된 문구 수정
