--- src/main/java/com/fittura/domain/product/categry/dto/response/CategoryResDto.java
@@ -1,24 +1,0 @@
-package com.fittura.domain.product.categry.dto.response;
-
-import com.fittura.domain.product.categry.constant.CategoryStatus;
-import com.fittura.domain.product.categry.entity.Category;
-
-public record CategoryResDto(
-    Long id,
-    String name,
-    Long parentId,
-    int depth,
-    int sortOrder,
-    CategoryStatus status
-) {
-    public static CategoryResDto from(Category category) {
-        return new CategoryResDto(
-            category.getId(),
-            category.getName(),
-            category.getParent() == null ? null : category.getParent().getId(),
-            category.getDepth(),
-            category.getSortOrder(),
-            category.getStatus()
-        );
-    }
-}
