--- src/main/java/com/fittura/global/webmvc/WebMvcConfig.java
@@ -1,27 +1,0 @@
-package com.fittura.global.webmvc;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-    @Value("${cors.allowed-origins}")
-    private String[] allowedOrigins;
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        for(String origin : allowedOrigins) {
-            if ("*".equals(origin)) {
-                throw new IllegalStateException("CORS 설정 오류: allowCredentials(true)와 함께 '*' 오리진은 사용할 수 없습니다. 구체적인 프론트 도메인을 설정하세요.");
-            }
-        }
-        registry.addMapping("/**")
-                .allowedOrigins(allowedOrigins)
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS", "HEAD")
-                .allowedHeaders("*")
-                .allowCredentials(true)
-                .maxAge(3600);
-    }
-}
